school/requirements.txt
cryptography==41.0.7
lxml==5.1.0
PyExecJS==1.5.1
requests==2.31.0
pydantic==2.6.3
pydantic_core==2.16.3
